Make middleware api route group specific
Rename the `Redaktor` middleware to make it message(request/response) specific.
Instead of having a single global middleware that runs on every single request,
potentially slowing down "web" requests in hybrid applications, we are making
the middleware "api" specific, so only api routes are revised (if applicable).

The downside of this approach is that since at this point the request has been
dispatched to the router and a route has been matched, it is a bit too late in
the request lifecycle to modify the request's URI. This means that this
middleware will only execute request and response revisions, leaving routes
revisions to be run by a global middleware that will be implemented at a later
point.