Expand route tagging support to route groups
Route group tagging is supported by passing a `tags` key in the `$attributes`
array passed in to the `group()` method. The value of the `tags` array element
must be a string or an array of strings.

Note: Group tagging does not perform any validation. This is due to the lack of
control when setting the `tags` attribute.