Add support for multi-strategy version resolver
This changes will allow us to use multiple strategies to resolve a version.

The current `VersionResolver` uses the configured strategies in a sequential
order to try resolve a version for the given request. If a version cannot be
resolved using a strategy, the next configured strategy will be used. In the
case that a version has not been resolved, an `UnresolvedVersionException` will
be thrown.