Add migration for version persistence
It gives us a base to implement the Database strategy. The Database strategy
will fetch a previously stored version from a database for a given identifier,
avoiding integrators having to specify the target version for every single
request. Additionally, it will allow devs using Redaktør to pin integrators to
a specific version so their integration is not broken if they're not explicitly
indicating the target version and a new version is released.