Implement `DatabaseStrategy` strategy
The `DatabaseStrategy` strategy resolves versions from the database based on a
given identifier (`Application-Id` header by default). That is, instead of an
application having to send the version for every single request, a version can
be _pinned_ (stored in the database) to a given unique identifier used in every
request (e.g. an application or client id). From then on, every subsequent
request with the given identifier will be resolved to the _pinned_ version.

A potential usage for this strategy is to pin all integrators to the latest
version before releasing a new one; that way, even if an integrator is not
sending the target version on every request, Redaktør will resolve their
_pinned_ version without breaking the integration.